schema-derive: refactor
